requirements.txt
langchain-core>=1.2.0
langchain-groq>=1.1.0
langgraph>=1.1.0
gitpython>=3.1.40
typer>=0.12.0
python-dotenv>=1.0.0
rich>=13.0.0

#for  local llm
langchain-ollama